+++ CHANGELOG.md
@@ -2,22 +2,48 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+Log of unreleased changes.
+
+### Added
+
+- Added Flickity Views style, it renders Views rows in a Flickity carousel.
+
+### Changed
+
+- Refactored `CHANGELOG.md` using *Keepachangelog.com* advices.
+
 ## v0.1.1
 
-- Typo, README links.
+Released on **2017/03/23**.
 
+### Fixed
+
+- Fixed typo.
 - Fixed Drupal is not defined for anonymous user.
 
 ## v0.1.0
 
-- Downloaded Flickity v2.0.5, added libraries.yml file. Flickity config entity.
+Released on **2017/02/23**.
 
-- Advanced options in Yaml format. Flickity field formatter.
+### Added
 
-- Fixed getFormattedOptions(). Attached library to field. Implemented
-flickityInit. Improved options form.
+- Added libraries.yml file.
+- Flickity config entity.
+- Added Advanced options in Yaml format.
+- Implemented Flickity field formatter.
+- Added `flickityInit.js` file.
+- Added cacheability metadata.
+- Added description and README.
+- Added CSS classes.
 
-- Cacheability metadata. Move Flickity library to libraries/ folder. Fixed
-adaptiveHeight option.
+### Changed
 
-- Description and README. Added CSS classes.
+- Improved options form.
+- Move Flickity library to `libraries/` folder.
+
+### Fixed
+
+- Fixed getFormattedOptions().
+- Fixed adaptiveHeight option.